Only allow school owners to create schools
As part of the verification flow, the user should have accepted the
responsibilities of a school owner by the time they come to create the
school. We need make sure we assign the school-owner role to them at
this point, which would also require the organisation to be created so
that we can assign the organisation_id on the Assignment in profile.