Create organisations via the ProfileApiClient
Although this API doesn’t exist yet, call the client in the relevant
parts of the code so that we would assign the organisation_id to the
school that is being created.

I think it makes sense for the creation of the organisation to happen
after we’ve sense checked that the school is valid, to avoid creating
unnecessary organisations in the profile app. We could create the
organisations earlier in the flow and then enforce the presence of
the ‘school-owner’ role via CanCanCan but I think it’s better to try and
make this as transactional as possible.