Add support for creating projects within a school’s library
If the user has the school-owner or school-teacher role, they can create
a project that is associated with a school (but not a school class).

If the user is a school-owner they can assign a teacher to the project
in the same was as lessons, otherwise the school-teacher is always set
as the project’s owner (user).