Improvements to User.from_omniauth spec setup
Previously this spec was making use of a `User` model and the
`User#serializable_hash` method. However, I'm pretty sure that was just
a convenience and there is nothing actually tying the attributes
returned from OmniAuth to the `User` model attributes.

I've changed the setup so it uses an explicit `Hash` derived from
observing the attributes that OmniAuth actually provides. This seems
more realistic and I'm hoping it will make it easier to apply an
unrelated fix in a subsequent commit.

I've also changed the assertions to be more explicit and like those
in the specs for `User.from_userinfo` & `User.from_token`.