Enforce Naming/VariableNumber cop with snake_case
Rather than disabling the Naming/VariableNumber cop [1] in a bunch of
places in the code, it seems simpler and more consistent to enable
t he cop globally, but configure it to use the `snake_case` option.

This assumes that `address_line_1` & `address_line_2` are using the
style that we want to use for *any* such symbols, variables or method
names. I've changed the lesson variables in the School spec to respect
this style. I've also added some exceptions using the `AllowedIdentifiers`
configuration option.

[1]: https://docs.rubocop.org/rubocop/cops_naming.html#namingvariablenumber