Do not compile assets in production
We're seeing the following warning in deployments like this one [1]:

    You set your `config.assets.compile = true` in production.
    This can negatively impact the performance of your application.

    For more information can be found in this article:
      https://devcenter.heroku.com/articles/rails-asset-pipeline#compile-set-to-true-in-production

Heroku is automatically running `rake assets:precompile` and `rake
assets:clean` (which just removes old versions of assets) as part of the
build step and the build will fail if either of these tasks fails. So we
should be able to rely on precompiled assets being present and thus we
should be able to set `config.assets.compile` in the Rails production
environment.

[1]: https://dashboard.heroku.com/apps/editor-api-staging/activity/builds/f9360560-9d26-487b-8258-cf84906055bf