Extract PhraseIdentifier.seed! method
This method populates data that is needed by the
`PhraseIdentifier.generate` method and so I think it makes things a bit
clearer to move the code onto the `PhraseIdentifier` class.

I've used the `chomp` option for `File.readlines` to simplify the code
slightly.

I've also added a call to `Word.delete_all` in order to make the method
idempotent which seems sensible given it can be called from `rails
db:seed` multiple times and I don't think we'd want to add the same
words over and over again.

Another motivation for this change is that I want to be able to call
`PhraseIdentifier.seed!` from a spec.