Change task spec to test FilesystemProject.import_all!
This seems to fix the following exception we were seeing when running
the spec on its own and within the suite (depending on the order specs
were run in):

    NameError: Rails couldn't find a valid model for ActiveStorage::Blob association

Testing the code outside the context of a Rake task is simpler and
doesn't really sacrifice much of the test coverage.

Previously even though the spec was looping through all the starter
projects it was stubbing the project config YAML and the contents of all
files. I found this quite confusing and so I've removed the stubbing and
the test doesn't take that long to run (3.5 seconds on my machine
excluding loading time). So I think making the spec more realistic like
this is worthwhile. If it's taking too long, we can look to change the
spec to only load a subset of the projects.