Fix order of schools.reference in schema.rb
In #267, this column was removed and then re-added. However, the merge
commit [1] only included a single migration that did not add/remove the
column was included, i.e. there was no net effect on
`schools.reference`, but the `reference` column was moved from near the
top to near the bottom of the table in `schema.rb`.

Ideally `schema.rb` should reflect the state of the production database
and so I have moved `reference` column back to where it was.

[1]: f3248986273f62017a8d0bef58bdabc6282c6d78