Remove calls to `ProfileApiClient` from `School::Create`
The `ProfileApiClient` was built to provide an interface to an imagined
future Profile API. We don't currently need this Profile API to be able
to make progress with the work of school onboarding in Editor Standalone
so we're removing it from school creation for now. This means that
Editor API will be the canonical source of Schools.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>