Remove code stubbing Profile organisations that don't exist
The Profile API doesn't currently return the `organisations` object
with the `<organistion_id>: <role>` pairs. We've decided instead to
model that inside of the editor-api codebase.

This commit removes the faking of that data inside User and the
stubbed responses in the fixture files. This causes a number of tests
to break - hopefully fixing those tests by adding the modelling in
this codebase instead will get us to where we want to be.