Fallback to Referer for Project#remix_origin
When running Cypress E2E specs (which run in the Electron browser by
default), there is no "Origin" header on the requests and so the
"validation" for `Project#remix_origin` in
`Project::CreateRemix#validate_params` [1] fails, making it impossible
to remix a project.

The Electron browser does seem to provide a "Referer" header on the
requests, so I've changed the code to fallback to use that. I've used
`URI#origin` to format `Project#remix_origin` consistently, because
unlike the "Origin" header, the "Referer" header can include a path, a
query string & a fragment.

[1]: https://github.com/RaspberryPiFoundation/editor-api/blob/1bb7e7f74b8021418503bcf3af201844c130aa9a/lib/concepts/project/operations/create_remix.rb#L22-L25