Use words directly from file vs caching in db
There are less than 1000 rows in the file which means it really doesn't
take long to load and parse, so I don't really see the point in storing
them in the database.

Previously the app didn't work unless we'd seeded the database which was
an extra complication for developers and deployments. This change
removes that complication.

My actual motivation for making the change is to make the deployment of
a new instance of the app at test-editor-api.raspberrypi.org simpler.