More explicit stubbing of UserInfo API
The previous implementation of `stub_user_info_api` would respond to all
requests to the userinfo API and either return the user object in
users.json with an ID matching that from the request, or return an empty
array when the ID in the request didn't match any of the records in
users.json. We've split this out into 5 separate methods so that:

- we can make the tests more explicit by using appropriately named
  methods.
- we can later remove the hardcoded user IDs from the factories.

Note that the original implementation could handle the case where
multiple user IDs were sent in the request but it turns out we're not
relying on this in any of our tests.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>