Pass user_id to `stub_user_info_api_for_unknown_users`
So that the implementation mirrors that of
`stub_user_info_api_for_[owner|teacher|student]` which will allow us to
remove some duplication.

This also ensures we're not stubbing more than necessary for our tests
to pass.

Note that we've disabled the RSpec/Examplelength RuboCop rule for a
number of specs. We're hoping that with some additional changes we're
planning we'll be able to remove these temporary exceptions.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>