Remove hardcoded id from school factory
Note the addition of `params[:lesson][:school_id] = school.id` in one of
the tests in creating_a_lesson_spec.rb. This is to satisfy
`LessonsController#verify_school_class_belongs_to_school`. Without this
change `LessonsController#school` returns the school created in the
`let` in the context, and not the school created in this test.