Update `User#school_teacher?` to use `Role`
I'm creating the Roles in `stub_user_info_api_for_teacher` and
`authenticate_as_school_teacher` to avoid having to update all the tests
but I might reconsider this approach in future as I'm not 100% convinced
by it.

In the tests that update a `School#id` I first have to remove the
associated role to avoid violating the database FK constraint.

I have to check for the existence of a school in `#create_teacher_role`
as some of the calls to either `stub_user_info_api_for_teacher` or
`authenticate_as_school_teacher` pass a random UUID that doesn't
correspond to a School in the database. I intend to investigate this
separately as it feels as though our tests might not be as robust as
they could be.