Update `User#school_student?` to use `Role`
I'm creating the Roles in `stub_user_info_api_for_student` and
`authenticate_as_school_student` to avoid having to update all the tests
but I might reconsider this approach in future as I'm not 100% convinced
by it.

In the test that updates a `School#id` I first have to remove the
associated role to avoid violating the database FK constraint.

In the test that checks for the number of `School#roles` I first have to
delete the roles so that we're not including the role created by the
call to `stub_user_info_api_for_student` in the `before` block.