Ensure users can only have roles within a single school
I believe the API was designed with the assumption that a user could
have roles in multiple organisations. The frontend, however, has been
designed with the assumption that a user only has a role (or roles) in a
single school. Adding this validation intends to prevent us from ending
up with users having roles in multiple schools. We might want to relax
this in future if it turns out we want to modify the frontend to cope
with users having roles in multiple schools.