Avoid network request in ClassMember validation
I believe the network request to the UserInfo API (via `#with_student`)
was added on the assumption that we were going to model roles in the
Profile app. Now that roles are modelled in this app we no longer need
to connect to the UserInfo API in this validation.

Test "responds with nil attributes for students if the user profile
doesn't exist" doesn't need to change `ClassMember#student_id` in order
to simulate the student not existing in Hydra. It can do that with the
call to `stub_user_info_api_for_unknown_users` for the existing
`ClassMember#student_id`.

We have to update a number of tests to make them more realistic.
Previously we were assigning an arbitrary `SecureRandom.uuid` to
`ClassMember#student_id` in a number of tests. That worked because the
`#student_has_the_school_student_role_for_the_school` validation would
return true if there was no user returned from the User Info API.  The
validation now requires a `User` that has the student role for the
associated school.