Avoid network request in SchoolClass validation
I believe the network request to the UserInfo API (via `#with_teacher`)
was added on the assumption that we were going to model roles in the
Profile app. Now that roles are modelled in this app we no longer need
to connect to the UserInfo API in this validation.

We have to update a number of tests to make them more realistic.
Previously we were assigning an arbitrary `SecureRandom.uuid` to
`SchoolClass#teacher_id` in a number of tests. That worked because the
`#teacher_has_the_school_teacher_role_for_the_school` validation would
return true if there was no user returned from the User Info API. The
validation now requires a `User` that has the teacher role for the
associated school.

In "responds with nil attributes for teachers if the user profile
doesn't exist" and "responds with nil attributes for the teacher if
their user profile doesn't exist" we're simulating what happens when a
user identified by `SchoolClass#teacher_id` doesn't exist in the User
Profile API. We can do that by stubbing the User Profile API with the
existing `SchoolClass#teacher_id`, we don't need to assign a new value
to this ID.