Pass School ID to `#authenticate_as_school_owner`
Instead of random UUIDs that don't necessarily correspond to a School in
the database. This is in preparation for passing School instances to
`authenticate_as_school_owner` directly.

In all but one of the tests I'm creating the School in a `let` block. In
spec/graphql/mutations/remix_project_mutation_spec.rb I'm creating the
school in the call to `authenticate_as_school_owner` to avoid the
following RuboCop violation:

    RSpec/MultipleMemoizedHelpers: Example group has too many memoized
    helpers [9/8]