Simplify tests that login as different teacher
Each of these tests describes the behaviour when the logged in user is a
teacher in the school but _not_ the teacher that owns the school class
or lesson. The call to `authenticate_as_school_teacher(school:)` is
enough to simulate this without also having to change the
`SchooClass#teacher_id` or `Lesson#user_id` to a random value.