Move validation of email address into Invitation
It seems more idiomatic to rely on active record's validation than to
call `raise ArgumentError` inside
SchoolTeacher::Invite#invite_teacher.

I've used a custom validation method here to make it clearer in this
commit's diff how the new code relates to the old, but I'll replace
this with a more standard validation method subsequently.

Note the slight change in capitalisation in the error message due to
how active record formats error messages by default. I didn't think it
was important enough to preserve the existing behaviour so have gone
with the easier option.