Introduce InvitationMailer
This commit introduces a basic mailer to send an invitation to a
teacher. I've lifted the content, subject and `from` address of the
email from Figma[1]. For now, I've just added a plain text version of
the email as this feels like the simplest way. We can add a HTML
version later.

The email is currently missing a link to "join school". I'll tackle
that in a subsequent commit.

[1] https://www.figma.com/design/SrINjD1MCT6VWYf1Yc9fV1/Code-Editor---Designs?node-id=5769-4727&t=ju5fZFHYj336iFgR-0