Pass User instance to `#stub_user_info_api_for_owner`
This method became responsible for creating the owner roles in PR #298
but that was only really a shortcut to reduce the scope of that PR.  I
think it makes sense to create the roles outside of this method so that
this is only responsible for stubbing the UserInfo API.

The new `owner` factory instantiates a `User` and creates an "owner"
role associating the user with the school. If no school is supplied when
creating an `owner` then creating the `Role` will fail with a validation
error because the school is missing. This makes sense because the only
thing that makes a user an "owner" is the presence of the owner role
associating them with a school.