run rails app:update
As per the rails upgrade guide[1] I ran

    rails app:update

and then

    rubocop -A

to bring the code generated by rails app:update into line with our
style guide.

I discarded any changes that removed code we'd explicitly added to,
for example, config/application.rb.

[1] https://guides.rubyonrails.org/upgrading_ruby_on_rails.html

We hadn't set the `config.force_ssl` option to true back when the app
was created in 2b09a14aa. It feels like the kind of thing that might
cause an issue, so I've decided to leave it unset with this upgrade too.