Add uniqueness constraint to `School#creator_id`
At present we don't want a user to be able to create more than a single
school (although we have hopes of relaxing this constraint in future).

I've checked the staging and production databases for existing
violations of this constraint and it looks like we're OK: there are 5
schools in staging, each with a unique creator_id, and there's only a
single school in production.