Include a link to accept the invitation in the email
For the first time editor-api needs to know the URL of the
editor-standalone application in order to generate the URL to redeem
the invitation, so we've decided to follow the existing pattern and
store that in an ENV variable.

Note that the route `/invitations/<token>` doesn't currently exist in
editor-standalone, we'll need to add that later.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>