Configure email sending with Postmark
This follows the setup instructions for our `editor-api-staging` email
server[1]. We've changed the `from` address to match the one
configured in Postmark.

In development we're only enabling sending with Postmark if the
POSTMARK_API_TOKEN is set. In this case it's useful to see errors so
we also set `raise_delivery_errors` to true.

[1] https://account.postmarkapp.com/servers/13473469/streams/outbound/get-started

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>