Add tests for `Admin::SchoolsController`
I want to make sure that verifying and rejecting schools continues to
work as I make changes to the functionality.

Unfortunately we now see the following deprecation warning in our tests:

> DEPRECATION WARNING: The option :class_name is deprecated.
> Administrate should detect it automatically.

This is caused by our use of `Field::BelongsTo` to associate the school
and user in the `SchoolDashboard` configuration[1]. The `School#creator`
method isn't actually a `belongs_to` association so we have to pass the
`class_name` to let Administrate know that the method returns a user. I
can't see a way of avoiding this deprecation warning at the moment.

[1]: https://github.com/RaspberryPiFoundation/editor-api/blob/6a8f7903112de73b018d4b50f7701f0e0bc40924/app/dashboards/school_dashboard.rb#L14