Extract `School#reject`
To complement `School#verify!`. Note that this doesn't need to be a bang
method because the return value of calling `update` is passed through
`SchoolVerificationService#reject` and back to
`Admin::SchoolsController#reject`. The `School#verify!` method needs to
raise on error so that it triggers a transaction rollback in
`SchoolVerificationService#verify`.