Fix bugs in `SchoolVerificationService`
We were previously ignoring the result of the call to `School#update`
which meant that if the update failed we'd still create the owner and
teacher roles, and return true from `#verify`. Calling
`School#update!` means that if the update fails then a
`ActiveRecord::RecordInvalid` exception will be raised, the transaction
will be rolled back and the `rescue` block will be invoked.

Fixing the first problem revealed that the `rescue` block was attempting
to access `school_id` which doesn't exist.