Add Api::TeacherInvitationsController#show action
When the user clicks on the link in the teacher invitation email, they
will be taken to a page which allows the user to accept the invitation.
This page needs to display the name of the school to which they've been
invited.

Also we need to check that the invitation token is valid for the
currently logged-in user - these scenarios as handled with HTTP status
codes as follows:

* User is not logged in: 401 Unauthorized
* Token does not exist: 404 Not Found
* Token has expired: 403 Forbidden
* Invitation email has changed since token was created: 403 Forbidden
* Invitation email does not match current user email: 403 Forbidden
  * This response includes an error message to distinguish it from
    the other two 403 Forbidden responses