Add Api::TeacherInvitationsController#accept action
This is intended to handle when the teacher clicks the button to accept
the invitation. We need to do all the same checks we did for the `#show`
action, i.e. that the invitation token is valid for the currently
logged-in user.

When the action is successful, the currently logged-in user is assigned
the "teacher" role in the school associated with the invitation. The
`Invitation#accepted_at` timestamp is also updated with the current
time.

There are a number of validation errors that can cause the action to
fail. These result in a response with a 422 Unprocessable entity HTTP
status code with the validation error Hash from `Invitation#errors`
included in the "error" key of the JSON in the response body.