Rename Invitation -> TeacherInvitation
To match the two `/teacher_invitations` API endpoints added in previous
commits.

The invitation is already kind of specific to teachers and in any case,
I think it will be simpler to extract duplication if/when we come to add
other invitations, e.g. for school owners, rather than try to come up
with a generic invitation model upfront.

Note that I've left the `InvitationMailer` & `InvitationPreview`
unchanged, because they both already have `#invite_teacher` methods
which are already sufficient to discriminate between different types of
invitation.