Create school in Profile API when verified
Any failure to create the school using the Profile API will result in
verification failing. This is because we don't currently have any other
way of keeping the data between Editor API and Profile API in sync so
it's safest to fail fast if any part of the verification fails.