Install libvips for ActiveStorage
To avoid the following errors when viewing project images in /admin.

```
/usr/local/bundle/gems/ffi-1.16.3/lib/ffi/dynamic_library.rb:65:in
`load_library': Could not open library 'vips.so.42': vips.so.42: cannot
open shared object file: No such file or directory. (LoadError)

Could not open library 'libvips.so.42':
libvips.so.42: cannot open shared object file: No such file or
directory.
```

It looks as though vips became the default in Rails 7.0[1] so it's possible
this problem predates the upgrade to Rails 7.1 but I'm not sure.

[1]: https://guides.rubyonrails.org/upgrading_ruby_on_rails.html#active-storage-default-variant-processor-changed-to-vips