Add `SchoolCodeGenerator`
In preparation for using this to generate 6 digit codes for schools.

Note that `Random#rand` "returns a random integer greater than or equal
to zero and less than max"[1]. By setting `MAX_CODE` to 1,000,000 we're
asking for a random number between 0 and 999,999.

The code generated is in the format 'nn-nn-nn', where 'n' is a number
between 0 and 9.

[1]: https://ruby-doc.org/3.2.3/Random.html#method-i-rand