Add `School#code`
The intention is to use this human-friendly code to scope student logins
to a specific school in Profile.

The validation for `School#code` ensures that it must:

- be unique
- be present if the school is verified
- be absent if the school is not verified
- be in the format /\d\d-\d\d-\d\d/
- not be changed once it's set

The code is automatically set on the school in the `School#verify!`
method. This method will retry 5 times if it attempts to save a code
that already exists on another school.