Avoid using users.json in UserInfoApiClient
We're hoping to remove the `user.json` fixture file. It doesn't feel if
we lose anything by hard-coding the user data like this for the bypass
oauth scenario. In particular, it looks as if the `users.json` data
envisaged each of the users having different roles supplied by the user
info API. However, we're now managing user roles within editor-api.