Avoid defaulting env vars in `AuthController#destroy`
I _think_ the only time it's safe to set these to nil is if BYPASS_OAUTH
is set to true (i.e. when we're not logged in via Profile). In that case
this `#destroy` method returns early so we can assume that they'll be
set if we're logging out for real.