Configure ActiveRecord encryption
Following the ActiveRecord encryption rails guide[1] this commit
configures active record with the necessary encryption keys and salts.

In development the values of these variables doesn't matter, but to
avoid confusion I've set them to something that isn't obviously an
encryption key (such as the ones generated by
`db:encryption:init`). We'll need to ensure that these three
environment variables are set in non-local environments before this
change is deployed.

[1] https://guides.rubyonrails.org/active_record_encryption.html