Encrypt the email_address field on TeacherInvitation
As a general rule we'd try to avoid having personally identifiable
information in the editor-api database. At the moment however, the
feature to generate email invitations has to live in this codebase. To
improve the security of the email_address field, this commit ensures
that is encrypted against the keys introduced in the previous commit.

I thought I might have to use the `deterministic: true` option[1] here
to allow the `find_by_token_for` method to work, but I've convinced
myself using the existing tests and by experimenting in the console
that this isn't the case. The generated token can be used to retrieve
the TeacherInvitation record even when `email_address` is encrypted.

I've added a test to demonstrate that this non-deterministic
encryption (that is, where a different value is generated each time
encrypt is called even for the same input) is intended. The rails
guide suggests that this is the best practice unless we need to query
by email address (which we currently don't need to do).

[1] https://guides.rubyonrails.org/active_record_encryption.html#deterministic-and-non-deterministic-encryption