Add `ProfileApiClient.delete_safeguarding_flag`
We're not using this at the moment but it makes it easier to test in
development if we can create and delete safeguarding flags for a user.

Note that I'm intentionally not doing anything different if
`BYPASS_OAUTH` is set. Given how tightly coupled Editor API and Profile
API are I'm less convinced it makes sense to attempt to run one without
the other in development.

API docs: https://my.raspberrypi.org/api/v1/documentation/#/Safeguarding%20Flags/delete_safeguarding_flags__flag_