Implement `ProfileApiClient.create_school_student`
Note that if the student is successfully created in Profile API then we
receive their ID in the response. We use this to create the relevant
student role and also make it available in the OperationResponse
returned by `SchoolStudent::Create.call` so that clients of Editor API
can use it if necessary.

I'm stripping any extraneous leading and trailing spaces from name,
username and password in ProfileApiClient.create_school_student to avoid
the sorts of problems that can occur when usernames and passwords
include these unnecessary spaces. I've tested the Profile API and it's
not (currently, at least) doing any kind of normalizing of what we pass
to it.

Note that I'm intentionally not doing anything different if
`BYPASS_OAUTH` is set. Given how tightly coupled Editor API and Profile
API are I'm less convinced it makes sense to attempt to run one without
the other in development.

API docs: https://my.raspberrypi.org/api/v1/documentation/#/Students/post_schools__schoolId__students