Extract `ProfileApiClient::SafeguardingFlag`
I find it easier to reason about the code when we're dealing in concrete
objects rather than hashes whose values are entirely dependent on the
response from an API. This implementation has the advantage of failing
fast if the API changes how it returns safeguarding flags. We use Ruby's
Data class[1] as this is a simple value object.

[1]: https://docs.ruby-lang.org/en/master/Data.html