Allow for dynamic hosts in CORS origin config (#409)
## Status

- Closes https://github.com/RaspberryPiFoundation/editor-api/issues/405

## What's changed?

This extends the current origin-setting functionality (literal origins
set via a comma-separated environment variable) and allows the env var
string to also contain regular expressions. These allow the origins to
match dynamic patterns (eg. `https://*.projects-ui.pages.dev`) and
ensures that configuration still remains in the environment as we don't
want to assume that anyone running this application will want the same
allowed origins configured as us (other than for `localhost` for local
and test environments only).

eg. instead of currently having to set the following in
`ALLOWED_ORIGINS`:

```
https://foo.raspberrypi.org, https://bar.raspberrypi.org, https://foo.editor-standalone-eyq.pages.dev, https://bar.editor-standalone-eyq.pages.dev
```

The following can be set:

```
/https:\/\/(?:[a-z0-9-]+\.)?raspberrypi\.org$/, /https:\/\/.+\.editor-standalone-eyq\.pages\.dev$/
```

Associated updates in Terraform allow for the origins to match the spec
in the
[issue](https://github.com/RaspberryPiFoundation/editor-api/issues/405)
(eg. the addition of wildcards / regex matching):

* https://github.com/RaspberryPiFoundation/terraform/pull/898